cleanup: split card into multiple components
