+++ src/app/home/page.js
@@ -1,6 +1,6 @@
 /**
  * Portfolio
- * Copyright (C) 2024 Maxim (https://github.com/max1mde)
+ * Copyright (C) 2024 Maxim (https://github.com/max1mde/portfolio)
  * 
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
