+++ src/app/privacy/page.js
@@ -1,7 +1,7 @@
 /**
  * Portfolio
- * Copyright (C) 2024 Maxim (https://github.com/max1mde)
- * 
+ * Copyright (C) 2024 Maxim (https://github.com/max1mde/portfolio)
+ *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
  * by the Free Software Foundation.
